Factor out _symbols and JSON functions to separate "api" module
