+++ src/pmt.c
@@ -26,6 +26,8 @@
 #include "socketworks.h"
 #include "tables.h"
 #include "utils.h"
+#include "api/symbols.h"
+#include "api/variables.h"
 #include "utils/ticks.h"
 
 #include <arpa/inet.h>
