+++ src/stream.c
@@ -43,6 +43,8 @@
 #include "stream.h"
 #include "t2mi.h"
 #include "pmt.h"
+#include "api/symbols.h"
+#include "api/variables.h"
 #include "utils/ticks.h"
 
 #define DEFAULT_LOG LOG_STREAM
